requirements.txt
Babel==2.3.4
backports.shutil-get-terminal-size==1.0.0
boto==2.40.0
boto3==1.3.1
botocore==1.4.18
decorator==4.0.9
defusedxml==0.4.1
Django==1.9.5
django-allauth==0.25.2
django-autofixture==0.12.1
django-bootstrap3==7.0.1
django-countries==3.4.1
django-extensions==1.6.1
django-fontawesome==0.3.1
django-localflavor==1.5
django-mailbox==4.5.4
django-phonenumber-field==1.1.0
django-storages==1.4.1
docutils==0.12
httplib2==0.9.2
ipython==4.2.0
ipython-genutils==0.1.0
jmespath==0.9.0
oauth2==1.9.0.post1
oauthlib==1.0.3
pexpect==4.0.1
phonenumberslite==7.3.1
pickleshare==0.7.2
#Pillow==3.2.0
ptyprocess==0.5.1
PyJWT==1.4.0
python-dateutil==2.5.3
python-docx==0.8.6
python-http-client==1.2.3
python-openid==2.2.5
python3-openid==3.0.10
pytz==2016.3
requests==2.9.1
requests-oauthlib==0.6.1
sendgrid==2.2.1
sendgrid-django==2.0.0
shortuuid==0.4.3
simplegeneric==0.8.1
six==1.10.0
smtpapi==0.3.1
traitlets==4.2.1
